Renamed plugin from schrodingerScipion to schrodinger
