+++ README.rst
@@ -32,24 +32,24 @@ OR
 
 .. code-block::
 
-            git clone https://github.com/scipion-chem/scipion-chem-schrodingerScipion.git
+            git clone https://github.com/scipion-chem/scipion-chem-schrodinger.git
 
 2. **Switch to the desired branch** (master or devel):
 
-Scipion-chem-schrodingerScipion is constantly under development and including new features.
+Scipion-chem-schrodinger is constantly under development and including new features.
 If you want a relatively older an more stable version, use master branch (default).
 If you want the latest changes and developments, user devel branch.
 
 .. code-block::
 
-            cd scipion-chem-schrodingerScipion
+            cd scipion-chem-schrodinger
             git checkout devel
 
 3. **Install**:
 
 .. code-block::
 
-            scipion3 installp -p path_to_scipion-chem-schrodingerScipion --devel
+            scipion3 installp -p path_to_scipion-chem-schrodinger --devel
 
 - **Binary files** 
 
@@ -63,7 +63,7 @@ or define the Schrodinger path in scipion.conf as SCHRODINGER_HOME
 
 To check the installation, simply run the following Scipion test:
 
-scipion3 tests schrodingerScipion.tests.test_sitemap.TestSitemap
+scipion3 tests schrodinger.tests.test_sitemap.TestSitemap
 
 ===============
 Buildbot status
