+++ schrodinger/tests/test_quikprop.py
